refactor: extract executor
